+++ tests/testthat/test-fct_drop.R
@@ -13,3 +13,12 @@ test_that("can optionally restrict levels to drop", {
   expect_equal(levels(fct_drop(f1, only = "b")), c("a", "c"))
   expect_equal(levels(fct_drop(f1, only = "a")), c("a", "b", "c"))
 })
+
+test_that("preserves ordered class and attributes", {
+  f1 <- ordered(letters[1:2], letters[1:3])
+  attr(f1, "x") <- "test"
+
+  f2 <- fct_drop(f1)
+  expect_s3_class(f2, "ordered")
+  expect_equal(attr(f2, "x"), attr(f1, "x"))
+})